Added in an ArrayList of Icons
There are two ArrayList of icons that have been added to the main. yourWinnings keeps track of the winning cards for the player, computerWinnings keeps track of the winnings for the computer. I don't quite know how to display the winnings to the GUI but I don't think we need to.